Wallet - keystore refactor
