+++ framework/snippets/src/imports.rs
@@ -5,9 +5,7 @@ pub use crate::{
     InteractorSimulateGasAsync, SimulateGas, StepBuffer, dns_address_for_name,
 };
 
-pub use crate::sdk::{
-    data::keystore::InsertPassword, test_wallets, validator::Validator, wallet::Wallet,
-};
+pub use crate::sdk::{test_wallets, validator::Validator, wallet::InsertPassword, wallet::Wallet};
 
 pub use env_logger;
