+++ sdk/core/src/data.rs
@@ -2,7 +2,6 @@ pub mod account;
 pub mod account_storage;
 pub mod esdt;
 pub mod hyperblock;
-pub mod keystore;
 pub mod network_config;
 pub mod network_economics;
 pub mod network_status;
